@@ -122,79 +122,82 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@graphiql/plugin-explorer": "^1.0.2",
-    "@mui/icons-material": "^5.14.16",
-    "@mui/lab": "^5.0.0-alpha.152",
-    "@mui/material": "^5.14.17",
-    "@mui/x-data-grid": "^6.18.0",
-    "@mui/x-date-pickers": "^6.18.0",
+    "@mui/icons-material": "^5.14.19",
+    "@mui/lab": "^5.0.0-alpha.155",
+    "@mui/material": "^5.14.20",
+    "@mui/x-data-grid": "^6.18.3",
+    "@mui/x-date-pickers": "^6.18.3",
     "@mui/x-tree-view": "^6.17.0",
     "@prettier/plugin-xml": "^3.2.2",
-    "@reduxjs/toolkit": "^1.9.7",
+    "@reduxjs/toolkit": "^2.0.1",
     "@stomp/stompjs": "^7.0.0",
     "@types/ace": "^0.0.51",
     "@types/graphql": "^14.5.0",
     "@types/marked": "^4.3.2",
-    "@types/react": "^18.2.36",
-    "@types/react-dom": "^18.2.14",
-    "@types/react-swipeable-views": "^0.13.4",
-    "@types/video.js": "^7.3.55",
-    "@videojs/vhs-utils": "^4.0.0",
+    "@types/react": "^18.2.42",
+    "@types/react-dom": "^18.2.17",
+    "@types/react-swipeable-views": "^0.13.5",
+    "@types/video.js": "^7.3.56",
+    "@videojs/vhs-utils": "^4.1.0",
     "autosuggest-highlight": "^3.3.4",
     "clsx": "^2.0.0",
     "fast-xml-parser": "^4.3.2",
-    "graphiql": "^3.0.9",
+    "graphiql": "^3.0.10",
     "graphql": "^16.8.1",
     "graphql-ws": "^5.14.2",
     "highlight.js": "^11.9.0",
     "jabber": "^1.5.3",
     "js-cookie": "^3.0.5",
-    "marked": "^9.1.5",
-    "marked-highlight": "^2.0.6",
+    "marked": "^11.0.1",
+    "marked-highlight": "^2.0.9",
     "moment-timezone": "^0.5.43",
-    "nanoid": "^5.0.3",
+    "nanoid": "^5.0.4",
     "notistack": "3.0.1",
-    "prettier": "^3.0.3",
+    "prettier": "^3.1.0",
     "query-string": "^8.1.0",
     "react": "18.2.0",
     "react-animations": "^1.0.0",
     "react-dom": "18.2.0",
     "react-hotkeys-hook": "^4.4.1",
     "react-infinite-scroller": "^1.2.6",
-    "react-intl": "^6.5.2",
-    "react-redux": "^8.1.3",
+    "react-intl": "^6.5.5",
+    "react-redux": "^9.0.2",
     "react-router-dom": "^5.3.4",
     "react-scripts": "5.0.1",
     "react-swipeable-views": "^0.14.0",
     "react-swipeable-views-utils": "^0.14.0",
-    "redux": "^4.2.1",
+    "redux": "^5.0.0",
     "redux-observable": "^2.0.0",
     "rxjs": "^7.0.0",
     "slugify": "^1.6.6",
     "tss-react": "^4.9.3",
-    "typescript": "^4.9.5",
-    "uppy": "^3.18.1",
-    "video.js": "^8.6.1",
+    "typescript": "^5.3.3",
+    "uppy": "^3.20.0",
+    "video.js": "^8.7.0",
     "web-vitals": "^3.5.0",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
     "@craco/craco": "7.1.0",
-    "@formatjs/cli": "^6.2.2",
+    "@formatjs/cli": "^6.2.4",
     "@rollup/plugin-typescript": "^11.1.5",
     "@types/js-cookie": "^3.0.6",
     "@types/prettier": "^3.0.0",
-    "@types/react-infinite-scroller": "^1.2.4",
-    "@types/zxcvbn": "^4.4.3",
-    "babel-plugin-formatjs": "^10.5.8",
+    "@types/react-infinite-scroller": "^1.2.5",
+    "@types/zxcvbn": "^4.4.4",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "babel-plugin-formatjs": "^10.5.10",
     "babel-plugin-react-intl": "^8.2.25",
     "build-if-changed": "^1.5.5",
-    "eslint": "^8.53.0",
+    "eslint": "^8.55.0",
     "eslint-config-react-app": "^7.0.1",
+    "fork-ts-checker-webpack-plugin": "^6.5.3",
     "ncp": "^2.0.0",
     "npm-run-all": "^4.1.5",
     "postcss-normalize": "^10.0.1",
     "prop-types": "^15.8.1",
     "rimraf": "^5.0.5",
-    "rollup": "^3.29.4"
+    "rollup": "^4.7.0"
   }
 }
