@@ -101,33 +101,33 @@
     "@emotion/css": "^11.11.2",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.16",
-    "@mui/material": "^5.14.17",
-    "@reduxjs/toolkit": "^1.9.7",
-    "@types/react-dom": "^18.2.14",
+    "@mui/icons-material": "^5.14.19",
+    "@mui/material": "^5.14.20",
+    "@reduxjs/toolkit": "^2.0.1",
+    "@types/react-dom": "^18.2.17",
     "js-cookie": "^3.0.5",
     "prop-types": "^15.8.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hotkeys-hook": "^4.4.1",
     "react-is": "^18.2.0",
-    "react-redux": "^8.1.3",
-    "redux": "^4.2.1",
+    "react-redux": "^9.0.2",
+    "redux": "^5.0.0",
     "redux-observable": "^2.0.0",
     "rxjs": "^7.0.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.3.3",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "@types/jquery": "^3.5.28",
+    "@types/jquery": "^3.5.29",
     "@types/js-cookie": "^3.0.6",
-    "@types/prop-types": "^15.7.9",
-    "@types/react-is": "^18.2.3",
-    "@types/tinymce": "^4.6.7",
+    "@types/prop-types": "^15.7.11",
+    "@types/react-is": "^18.2.4",
+    "@types/tinymce": "^4.6.9",
     "build-if-changed": "^1.5.5",
     "npm-run-all": "^4.1.5",
     "replace-in-file": "^7.0.2",
-    "rollup": "^3.29.4",
+    "rollup": "^4.7.0",
     "yargs": "^17.7.2"
   }
 }
