@@ -34,12 +34,12 @@
     "watch": "babel src -d lib --watch"
   },
   "devDependencies": {
-    "@babel/cli": "^7.23.0",
-    "@babel/plugin-transform-object-assign": "^7.22.5",
-    "@types/eslint": "^8",
+    "@babel/cli": "^7.23.4",
+    "@babel/plugin-transform-object-assign": "^7.23.3",
+    "@types/eslint": "^8.44.8",
     "babel-plugin-inline-package-json": "^2.0.0",
     "build-if-changed": "^1.5.5",
-    "eslint": "^8.53.0"
+    "eslint": "^8.55.0"
   },
   "peerDependencies": {
     "uppy": "^2.13.8"
