@@ -14,10 +14,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { get, postJSON } from '../utils/ajax';
-import { pluck } from 'rxjs/operators';
+import { post, postJSON } from '../utils/ajax';
+import { map, pluck } from 'rxjs/operators';
 import { Observable } from 'rxjs';
-import { toQueryString } from '../utils/object';
 
 export function encrypt(text: string): Observable<string>;
 export function encrypt(text: string, site: string): Observable<string>;
@@ -25,9 +24,6 @@ export function encrypt(text: string, site: string = ''): Observable<string> {
   return postJSON(`/studio/api/2/security/encrypt.json`, { text, siteId: site }).pipe(pluck('response', 'item'));
 }
 
-export function getUserPermissions(site: string, path: string): Observable<string[]> {
-  const qs = toQueryString({ site, path });
-  return get(`/studio/api/1/services/api/1/security/get-user-permissions.json${qs}`).pipe(
-    pluck('response', 'permissions')
-  );
+export function previewSwitch(): Observable<true> {
+  return post(`/studio/api/2/security/preview/switch`).pipe(map(() => true));
 }
