@@ -45,10 +45,10 @@ export function NewRemoteRepositoryDialogContainer(props: NewRemoteRepositoryDia
         ...(inputs.repoAuthentication === 'basic'
           ? { remoteUsername: inputs.repoUsername, remotePassword: inputs.repoPassword }
           : inputs.repoAuthentication === 'token'
-          ? { remoteUsername: inputs.repoUsername, remoteToken: inputs.repoToken }
-          : inputs.repoAuthentication === 'key'
-          ? { remotePrivateKey: inputs.repoKey }
-          : {})
+            ? { remoteUsername: inputs.repoUsername, remoteToken: inputs.repoToken }
+            : inputs.repoAuthentication === 'key'
+              ? { remotePrivateKey: inputs.repoKey }
+              : {})
       }).subscribe({
         next: () => {
           functionRefs.current.onSubmittingAndOrPendingChange({
