@@ -120,7 +120,7 @@ export function RecentlyPublishedDashlet(props: RecentlyPublishedDashletProps) {
 
   // region Item Updates Propagation
   useEffect(() => {
-    const events = [publishEvent.type];
+    const events: string[] = [publishEvent.type];
     const hostToHost$ = getHostToHostBus();
     const subscription = hostToHost$.pipe(filter((e) => events.includes(e.type))).subscribe(({ type, payload }) => {
       loadPage(getCurrentPage(offset, limit), true);
