@@ -14,28 +14,25 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+export type Primitive = string | number | boolean | null | undefined | Primitive[] | { [key: string]: Primitive };
+
 export interface ContentInstanceSystemProps {
   id: string;
-  path: string;
+  path: string | null;
   label: string; // Internal name
   dateCreated: string;
   dateModified: string;
   contentTypeId: string;
-  sourceMap?: { [fieldId: string]: string }; // fieldId: path
+  sourceMap?: Record<string, string>; // { fieldId: path }
 }
 
-export interface ContentInstance {
+export interface ContentInstanceBase {
   craftercms: ContentInstanceSystemProps;
-  [fieldId: string]: any;
 }
+
+export type ContentInstance<T extends Record<string, any> = Record<string, any>> = T & ContentInstanceBase;
 
 // An InstanceRecord is a ContentInstance without ContentInstanceSystemProps
-export type InstanceRecord = Record<string, string | number | boolean | any[]>;
-
-// TODO: ContentInstance typing is too loose.
-//  Something like this may be more suitable. Some compilation issues would arise, though.
-//  export type ContentInstance = { craftercms: ContentInstanceSystemProps } & {
-//    [fieldId: string]: number | string | boolean | object | Array<object | string>;
-//  };
+export type InstanceRecord = Record<string, Primitive>;
 
 export default ContentInstance;
