@@ -72,7 +72,7 @@ import { fetchSiteConfig } from '../actions/configuration';
 import { getStoredShowToolsPanel } from '../../utils/state';
 import { closeToolsPanel, openToolsPanel } from '../actions/preview';
 import { getXSRFToken, setSiteCookie } from '../../utils/auth';
-import { changeSite, fetchSites } from '../actions/sites';
+import { changeSiteComplete, fetchSites } from '../actions/sites';
 import { closeConfirmDialog, showConfirmDialog } from '../actions/dialogs';
 import { defineMessages } from 'react-intl';
 import { createCustomDocumentEventListener } from '../../utils/dom';
@@ -113,18 +113,18 @@ const systemEpics: CrafterCMSEpic[] = [
                 showToolsPanel === null || state.preview.showToolsPanel === showToolsPanel
                   ? false
                   : state.preview.showToolsPanel
-                  ? closeToolsPanel()
-                  : openToolsPanel()
+                    ? closeToolsPanel()
+                    : openToolsPanel()
               ].filter(Boolean)
             : [])
         ];
       })
     ),
   // endregion
-  // region changeSite
+  // region changeSiteComplete
   (action$) =>
     action$.pipe(
-      ofType(changeSite.type),
+      ofType(changeSiteComplete.type),
       switchMap(({ payload: { nextSite } }) => [
         startPublishingStatusFetcher(),
         fetchSiteConfig(),
