@@ -20,7 +20,7 @@ import { initDashboardConfig } from '../actions/dashboard';
 import { deserialize, fromString } from '../../utils/xml';
 import { applyDeserializedXMLTransforms } from '../../utils/object';
 import { defineMessages } from 'react-intl';
-import { changeSite } from '../actions/sites';
+import { changeSiteComplete } from '../actions/sites';
 
 const messages = defineMessages({
   noUiConfigMessageTitle: {
@@ -35,7 +35,7 @@ const messages = defineMessages({
 
 const reducer = createReducer<GlobalState['dashboard']>(null, (builder) => {
   builder
-    .addCase(changeSite, () => null)
+    .addCase(changeSiteComplete, () => null)
     .addCase(initDashboardConfig, (state, { payload }) => {
       let dashboardConfig = {
         widgets: [
