@@ -17,7 +17,7 @@
 import { createReducer } from '@reduxjs/toolkit';
 import GlobalState from '../../models/GlobalState';
 import { changeCurrentUrl, goToLastPage, goToNextPage, guestCheckIn } from '../actions/preview';
-import { changeSite } from '../actions/sites';
+import { changeSiteComplete } from '../actions/sites';
 import { deleteContentEvent, emitSystemEvent } from '../actions/system';
 import { getPreviewURLFromPath } from '../../utils/path';
 
@@ -119,7 +119,7 @@ const reducer = createReducer<GlobalState['previewNavigation']>(
       // - Clear the back and forward stacks
       // - Clear the navigation type
       // - Update the current url
-      .addCase(changeSite, (state, { payload }) => ({
+      .addCase(changeSiteComplete, (state, { payload }) => ({
         ...state,
         historyBackStack: [],
         historyForwardStack: [],
