@@ -50,6 +50,7 @@ import {
   changeContentType,
   editContentTypeTemplate,
   editController,
+  EditFilePayload,
   editTemplate
 } from '../state/actions/misc';
 import {
@@ -120,8 +121,8 @@ import { fetchPublishingStatus } from '../state/actions/publishingStatus';
 import { Clipboard } from '../models/GlobalState';
 import { Dispatch } from 'redux';
 import SystemType from '../models/SystemType';
-import StandardAction from '../models/StandardAction';
 import { fetchItemVersions } from '../state/actions/versions';
+import StandardAction from '../models/StandardAction';
 
 export type ContextMenuOptionDescriptor<ID extends string = string> = {
   id: ID;
@@ -589,8 +590,8 @@ export const itemActionDispatcher = ({
             item.systemType === 'renderingTemplate'
               ? 'template'
               : item.systemType === 'script'
-              ? 'controller'
-              : 'asset';
+                ? 'controller'
+                : 'asset';
 
           dispatch(
             showRenameAssetDialog({
@@ -970,7 +971,10 @@ export const itemActionDispatcher = ({
   }
 };
 
-export function editControllerActionCreator(systemType: SystemType, contentTypeId: string): StandardAction {
+export function editControllerActionCreator(
+  systemType: SystemType,
+  contentTypeId: string
+): StandardAction<EditFilePayload> {
   return editController({
     path: getControllerPath(systemType),
     fileName: `${popPiece(contentTypeId, '/')}.groovy`,
